Revert "Revert "simplify API""
This reverts commit eb1d17dd6c64f6671513a889e1719f5c7e644991.
