+++ prompthub/__init__.py
@@ -1,0 +1,1 @@
+from prompthub.prompt import Prompt, from_json, from_yaml, fetch
